+++ src/java/manifail/RetriesExceeded.java
@@ -6,8 +6,13 @@ package manifail;
 public class RetriesExceeded extends Throwable {
     private static final long serialVersionUID = 1552386065262456541L;
     public final int retries;
-    public RetriesExceeded(int retries, Throwable cause) {
-        super(cause);
+    public final Object value;
+    public RetriesExceeded(int retries) {
         this.retries = retries;
+        this.value = null;
+    }
+    public RetriesExceeded(int retries, Object value) {
+        this.retries = retries;
+        this.value = value;
     }
 }